Fixed GFX display broken by UTFT implementation

+++ menu.h
@@ -166,8 +166,8 @@ www.r-site.net
     virtual void print(char ch)=0;
     virtual void print(const char *text)=0;
     virtual void println(const char *text="")=0;
-    virtual void print(int)=0;
-    virtual void println(int)=0;
+    virtual void print(unsigned long)=0;
+    virtual void println(unsigned long)=0;
     virtual void print(double)=0;
     virtual void println(double)=0;
     virtual void print(prompt &o,bool selected,int idx,int posY,int width);
